requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.26.4
azure-iot-device==3.0.0b2
azure-iot-hub==2.6.1
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
deprecation==2.1.0
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.2
Flask-SocketIO==5.3.4
frozenlist==1.3.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
importlib-resources==5.12.0
isodate==0.6.1
itsdangerous==2.1.2
janus==1.0.0
Jinja2==3.1.2
limits==3.4.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
msrest==0.7.1
multidict==6.0.4
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.1
paho-mqtt==1.6.1
Pygments==2.15.1
PySocks==1.7.1
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.30.0
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
rich==13.3.5
six==1.16.0
typing_extensions==4.5.0
uamqp==1.6.4
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.15.0
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
